# ⛑ make fastmap safe and add test case for hash seed and prime