--- a/fastmap.go
+++ b/fastmap.go
@@ -1,9 +1,11 @@
 package utils
 
-const fastStrMapSliceLimit = 10000
+const fastStrMapSliceLimit = 1229
+const fastStrMapSeed = 31
 
 type fastStrMap struct {
 	data [fastStrMapSliceLimit]string
+	ref  [fastStrMapSliceLimit]string
 }
 
 func NewFastMap(m map[string]string) *fastStrMap {
@@ -20,21 +22,23 @@ func (fm *fastStrMap) setMap(m map[string]string) {
 
 func (fm *fastStrMap) set(key, val string) {
 	fm.data[strToInt(key)] = val
+	fm.ref[strToInt(key)] = key
 }
 
 func (fm *fastStrMap) Get(key string) string {
 	i := strToInt(key)
-	if i > fastStrMapSliceLimit {
+	if i >= fastStrMapSliceLimit || len(key) != len(fm.ref[i]) || fm.ref[i] != key {
 		return ""
 	}
 	return fm.data[i]
 }
 
+// strToInt uses BKDRHash algorithm
 func strToInt(s string) int {
-	seed, hash := 131, 0
+	var hash int
 
 	for _, b := range s {
-		hash = hash*seed + int(b)
+		hash = hash*fastStrMapSeed + int(b)
 	}
 	return (hash & 0x7fffffff) % fastStrMapSliceLimit
 }