--- a/fastmap.go
+++ b/fastmap.go
@@ -27,7 +27,7 @@ func (fm *fastStrMap) set(key, val string) {
 
 func (fm *fastStrMap) Get(key string) string {
 	i := strToInt(key)
-	if len(key) != len(fm.ref[i]) || fm.ref[i] != key {
+	if fm.ref[i] != key {
 		return ""
 	}
 	return fm.data[i]