… fastmap_test.go