NEW:
```
Benchmark_GetMIME/fiber-8               10162006               117 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8               10228453               118 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8               10361821               120 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8               10091090               117 ns/op               0 B/op          0 allocs/op
```

OLD:
Benchmark_GetMIME/fiber-8                9390570               133 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8                9346348               134 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8                9446317               131 ns/op               0 B/op          0 allocs/op
Benchmark_GetMIME/fiber-8                9408157               130 ns/op               0 B/op          0 allocs/op
```